+++ src/SqlDsl.Core/SqlIntExpr.cs
@@ -1,0 +1,52 @@
+namespace SqlDsl.Core
+{
+ 
+
+    public class SqlIntAdd : SqlBinExpr<SqlInt>
+    {
+        public SqlExpr<SqlInt> Left { get; }
+        public SqlExpr<SqlInt> Right { get; }
+        public SqlIntAdd(SqlExpr<SqlInt> left, SqlExpr<SqlInt> right)
+        {
+            this.Left = left;
+            this.Right = right;
+        }
+        public void Deconstruct(out SqlExpr<SqlInt> left, out SqlExpr<SqlInt> right)
+        {
+            left = this.Left;
+            right = this.Right;
+        }
+    }
+
+    public class SqlIntMult : SqlBinExpr<SqlInt>
+    {
+        public SqlExpr<SqlInt> Left { get; }
+        public SqlExpr<SqlInt> Right { get; }
+        public SqlIntMult(SqlExpr<SqlInt> left, SqlExpr<SqlInt> right)
+        {
+            this.Left = left;
+            this.Right = right;
+        }
+        public void Deconstruct(out SqlExpr<SqlInt> left, out SqlExpr<SqlInt> right)
+        {
+            left = this.Left;
+            right = this.Right;
+        }
+    }
+
+    public class SqlIntValue : SqlExpr<SqlInt>
+    {
+        public int Value { get; private set; }
+        public SqlIntValue(int v)
+        {
+            this.Value = v;
+        }
+        public void Deconstruct(out int value)
+        {
+            value = this.Value;
+        }
+    }
+
+ 
+
+}
