+++ src/SqlDsl.Core/SqlStringExpr.cs
@@ -1,0 +1,16 @@
+namespace SqlDsl.Core
+{
+
+    public class SqlStringConcat : SqlBinExpr<SqlString>
+    {
+        public SqlExpr<SqlString> Left { get; }
+        public SqlExpr<SqlString> Right { get; }
+        public SqlStringConcat(SqlExpr<SqlString> left, SqlExpr<SqlString> right)
+        {
+            this.Left = left;
+            this.Right = right;
+        }
+    }
+
+
+}
